requirements.txt
cvxpy                             1.5.2
jupyter                           1.0.0
jupyter-contrib-nbextensions      0.7.0
jupyterlab                        4.2.2
matplotlib                        3.9.0
Mosek                             10.2.1
numpy                             2.0.0
pip                               24.1
scipy                             1.13.1
sdpa-python                       0.2.1
setuptools                        65.5.1
wheel                             0.38.4
